change product price type from float64 to int change order money type from float64 to int

+++ business/iface.go
@@ -18,13 +18,13 @@ type ICorp interface {
 type IOrder interface {
 	GetId() int
 	GetBid() string
-	GetDeductableMoney() float64
+	GetDeductableMoney() int
 }
 
 type IResource interface {
 	GetType() string //获得资源类型
-	GetDeductionMoney(deductableMoney float64) float64 //获得资源抵扣的金额
-	GetPrice() float64 //获得资源价格
+	GetDeductionMoney(deductableMoney int) int //获得资源抵扣的金额
+	GetPrice() int //获得资源价格
 	GetPostage() float64 //获得资源的运费
 	IsAllocated() bool //设置资源已申请成功
 	ResetAllocation() //设置资源未申请
